test: seperate Chat and Match test code
- 채팅과 매칭내역 테스트 코드 분리
