+++ sunsu-wedding/src/test/java/com/kakao/sunsuwedding/chat/ChatRepositoryTest.java
@@ -1,0 +1,7 @@
+package com.kakao.sunsuwedding.chat;
+
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+@DataJpaTest
+public class ChatRepositoryTest {
+}
