+++ sunsu-wedding/src/test/java/com/kakao/sunsuwedding/match/MatchRestControllerTest.java
@@ -36,139 +36,8 @@ public class MatchRestControllerTest {
     @Autowired
     private ObjectMapper om;
 
-    // ============ 채팅방 생성 테스트 ============
-    @DisplayName("채팅방 생성 성공 테스트")
-    @Test
-    @WithUserDetails("couple3@gmail.com")
-    public void match_create_success_test() throws Exception {
-        //given
-        MatchRequest.AddMatchDTO requestDTO = new MatchRequest.AddMatchDTO();
-        requestDTO.setPlannerId(1L);
-        String requestBody = om.writeValueAsString(requestDTO);
+    // ============ 견적서 전체(매칭) 확정 테스트 ============
 
-        //when
-        ResultActions result = mvc.perform(
-                MockMvcRequestBuilders
-                        .post("/chat")
-                        .content(requestBody)
-                        .contentType(MediaType.APPLICATION_JSON)
-        );
-
-        logResult(result);
-
-        // then
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.success").value("true"));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.response.chatId").value(7L));
-    }
-
-    @DisplayName("채팅방 생성 실패 테스트 - 이미 존재하는 매칭내역")
-    @Test
-    @WithUserDetails("couple@gmail.com")
-    public void match_create_fail_test() throws Exception {
-        //given
-        MatchRequest.AddMatchDTO requestDTO = new MatchRequest.AddMatchDTO();
-        requestDTO.setPlannerId(2L);
-        String requestBody = om.writeValueAsString(requestDTO);
-
-        //when
-        ResultActions result = mvc.perform(
-                MockMvcRequestBuilders
-                        .post("/chat")
-                        .content(requestBody)
-                        .contentType(MediaType.APPLICATION_JSON)
-        );
-
-        logResult(result);
-
-        // then
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.success").value("false"));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.message").value("이미 존재하는 매칭입니다."));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.status").value(400));
-    }
-
-    // ============ 채팅방 삭제 테스트 ============
-    @DisplayName("채팅방 삭제 성공 테스트")
-    @Test
-    @WithUserDetails("couple@gmail.com")
-    public void match_delete_success_test() throws Exception {
-        //given
-        Long matchId = 1L;
-
-        //when
-        ResultActions result = mvc.perform(
-                MockMvcRequestBuilders
-                        .delete("/chat?matchId=" + matchId)
-        );
-
-        logResult(result);
-
-        // then
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.success").value("true"));
-    }
-
-    @DisplayName("채팅방 삭제 실패 테스트 1 - 견적서 전체 확정 되지 않음")
-    @Test
-    @WithUserDetails("couple@gmail.com")
-    public void match_delete_fail_test1() throws Exception {
-        //given
-        Long matchId = 2L;
-
-        //when
-        ResultActions result = mvc.perform(
-                MockMvcRequestBuilders
-                        .delete("/chat?matchId=" + matchId)
-        );
-
-        logResult(result);
-
-        // then
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.success").value("false"));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.status").value(400));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.message").value("견적서 전체 확정을 해야 합니다."));
-
-    }
-
-    @DisplayName("채팅방 삭제 실패 테스트 2 - 본인의 채팅방이 아님")
-    @Test
-    @WithUserDetails("couple@gmail.com")
-    public void match_delete_fail_test2() throws Exception {
-        //given
-        Long matchId = 3L;
-
-        //when
-        ResultActions result = mvc.perform(
-                MockMvcRequestBuilders
-                        .delete("/chat?matchId=" + matchId)
-        );
-
-        logResult(result);
-
-        // then
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.success").value("false"));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.status").value(403));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.message").value("사용할 수 없는 기능입니다."));
-    }
-
-    @DisplayName("채팅방 삭제 실패 테스트 3 - 존재하지 않는 채팅방(매칭 내역)")
-    @Test
-    @WithUserDetails("couple@gmail.com")
-    public void match_delete_fail_test3() throws Exception {
-        //given
-        Long matchId = 10L;
-
-        //when
-        ResultActions result = mvc.perform(
-                MockMvcRequestBuilders
-                        .delete("/chat?matchId=" + matchId)
-        );
-
-        logResult(result);
-
-        // then
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.success").value("false"));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.status").value(404));
-        result.andExpect(MockMvcResultMatchers.jsonPath("$.error.message").value("매칭 내역을 찾을 수 없습니다."));
-    }
 
     private void logResult(ResultActions result) throws Exception {
         String responseBody = result.andReturn().getResponse().getContentAsString();
